+++ pkg/transforms/daemonset_test.go
@@ -12,9 +12,9 @@ func TestTransformDaemonSet(t *testing.T) {
 	node := transformDaemonSet(&d)
 
 	// Test only the fields that exist in daemonset - the common test will test the other bits
-	AssertEqual("available", node.Properties["available"], int32(1), t)
-	AssertEqual("current", node.Properties["current"], int32(1), t)
-	AssertEqual("desired", node.Properties["desired"], int32(1), t)
-	AssertEqual("ready", node.Properties["ready"], int32(1), t)
-	AssertEqual("updated", node.Properties["updated"], int32(1), t)
+	AssertEqual("available", node.Properties["available"], int64(1), t)
+	AssertEqual("current", node.Properties["current"], int64(1), t)
+	AssertEqual("desired", node.Properties["desired"], int64(1), t)
+	AssertEqual("ready", node.Properties["ready"], int64(1), t)
+	AssertEqual("updated", node.Properties["updated"], int64(1), t)
 }
