+++ pkg/transforms/replicaset_test.go
@@ -12,6 +12,6 @@ func TestTransformReplicaSet(t *testing.T) {
 	node := transformReplicaSet(&r)
 
 	// Test only the fields that exist in replica set - the common test will test the other bits
-	AssertEqual("current", node.Properties["current"], int32(1), t)
-	AssertEqual("desired", node.Properties["desired"], int32(1), t)
+	AssertEqual("current", node.Properties["current"], int64(1), t)
+	AssertEqual("desired", node.Properties["desired"], int64(1), t)
 }
