+++ pkg/transforms/statefulset.go
@@ -11,10 +11,10 @@ func transformStatefulSet(resource *v1.StatefulSet) Node {
 
 	// Extract the properties specific to this type
 	statefulSet.Properties["kind"] = "StatefulSet"
-	statefulSet.Properties["current"] = resource.Status.Replicas
-	statefulSet.Properties["desired"] = int32(0)
+	statefulSet.Properties["current"] = int64(resource.Status.Replicas)
+	statefulSet.Properties["desired"] = int64(0)
 	if resource.Spec.Replicas != nil {
-		statefulSet.Properties["desired"] = *resource.Spec.Replicas
+		statefulSet.Properties["desired"] = int64(*resource.Spec.Replicas)
 	}
 
 	return statefulSet
